tests/requirements_test.txt
pytest>=8.0.0
pytest-asyncio>=0.26.0
pytest-cov>=4.1.0
pytest-html>=4.1.0
pytest-mock>=3.11.0
pytest-xdist>=3.5.0
pytest-env>=1.0.0
pytest-timeout>=2.2.0
pytest-benchmark>=4.0.0
pytest-sugar>=0.9.7
pytest-clarity>=1.0.1
pytest-randomly>=3.12.0

# 异步和HTTP测试
aiohttp>=3.8.4
requests>=2.31.0
httpx>=0.24.0
websockets>=11.0.0

# 数据处理和验证
click>=8.1.3
PyYAML>=6.0
pydantic>=2.0.0
marshmallow>=3.19.0

# 覆盖率和报告
coverage>=7.2.0
coverage-badge>=1.1.0

# 数据库和消息队列客户端
nats-py>=2.6.0
clickhouse-driver>=0.2.5
redis>=4.5.0

# 容器化测试
docker>=6.1.0
testcontainers>=3.7.0

# 性能和负载测试
locust>=2.15.0
memory-profiler>=0.61.0
psutil>=5.9.0
py-spy>=0.3.14

# Mock和测试数据生成
factory-boy>=3.2.1
faker>=19.0.0
responses>=0.23.0
freezegun>=1.2.2

# 代码质量
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.4.0

# 测试报告和可视化
allure-pytest>=2.13.0
pytest-json-report>=1.5.0